+++ src/gtf/reader.rs
@@ -5,8 +5,6 @@ use std::io::{BufRead, BufReader};
 use std::path::Path;
 use std::str::FromStr;
 
-use log::debug;
-
 use crate::gtf::{GtfFeature, GtfRecord, GtfRecordsGroup};
 use crate::models::{Transcript, TranscriptRead, Transcripts};
 use crate::utils::errors::ParseGtfError;
@@ -24,7 +22,7 @@ use crate::utils::errors::ReadWriteError;
 /// use atg::models::TranscriptRead;
 ///
 /// // create a reader from the tests GTF file
-/// let reader = Reader::from_file("tests/data/tests.gtf");
+/// let reader = Reader::from_file("tests/data/example.gtf");
 /// assert_eq!(reader.is_ok(), true);
 ///
 /// // parse the GTF file
@@ -33,7 +31,7 @@ use crate::utils::errors::ReadWriteError;
 ///     .transcripts()
 ///     .unwrap();
 ///
-/// assert_eq!(transcripts.len(), 10);
+/// assert_eq!(transcripts.len(), 27);
 /// ```
 pub struct Reader<R> {
     inner: std::io::BufReader<R>,
@@ -52,7 +50,7 @@ impl Reader<File> {
     /// use atg::models::TranscriptRead;
     ///
     /// // create a reader from the tests GTF file
-    /// let reader = Reader::from_file("tests/data/tests.gtf");
+    /// let reader = Reader::from_file("tests/data/example.gtf");
     /// assert_eq!(reader.is_ok(), true);
     ///
     /// // parse the GTF file
@@ -61,7 +59,7 @@ impl Reader<File> {
     ///     .transcripts()
     ///     .unwrap();
     ///
-    /// assert_eq!(transcripts.len(), 10);
+    /// assert_eq!(transcripts.len(), 27);
     /// ```
     pub fn from_file<P: AsRef<Path>>(path: P) -> Result<Self, ReadWriteError> {
         match File::open(path.as_ref()) {
@@ -128,7 +126,7 @@ impl<R: std::io::Read> TranscriptRead for Reader<R> {
     /// use atg::models::TranscriptRead;
     ///
     /// // create a reader from the tests GTF file
-    /// let reader = Reader::from_file("tests/data/tests.gtf");
+    /// let reader = Reader::from_file("tests/data/example.gtf");
     ///
     /// // parse the GTF file
     /// let transcripts = reader
@@ -136,7 +134,7 @@ impl<R: std::io::Read> TranscriptRead for Reader<R> {
     ///     .transcripts()
     ///     .unwrap();
     ///
-    /// assert_eq!(transcripts.len(), 10);
+    /// assert_eq!(transcripts.len(), 27);
     /// ```
     fn transcripts(&mut self) -> Result<Transcripts, ReadWriteError> {
         let mut transcript_hashmap: HashMap<String, GtfRecordsGroup> = HashMap::new();
@@ -163,8 +161,6 @@ impl<R: std::io::Read> TranscriptRead for Reader<R> {
                 _ => {}
             }
         }
-
-        debug!("Finished reading GTF input");
 
         let mut res = Transcripts::with_capacity(transcript_hashmap.len());